Split web route pages into named components
